requirements/local.txt
-r ./base.txt

Werkzeug==3.0.1  # The comprehensive WSGI web application library, https://github.com/pallets/werkzeug
#ipdb==0.11  # Integration of IPython pdb, https://github.com/gotcha/ipdb
Sphinx==7.2.6  # https://github.com/sphinx-doc/sphinx
#psycopg2-binary==2.7.5  # https://github.com/psycopg/psycopg2
pre-commit
pre-commit-hooks

# Testing
# ------------------------------------------------------------------------------
#pytest==3.5.0  # https://github.com/pytest-dev/pytest
#pytest-sugar==0.9.1  # https://github.com/Frozenball/pytest-sugar

# Code quality
# ------------------------------------------------------------------------------
#flake8==3.5.0  # https://github.com/PyCQA/flake8
#coverage==4.5.1  # https://github.com/nedbat/coveragepy
black==22.*
isort==5.8.*
bandit==1.7.5
django-upgrade

# Django
# ------------------------------------------------------------------------------
#factory-boy==2.10.0  # https://github.com/FactoryBoy/factory_boy
#django-test-plus==1.0.22  # https://github.com/revsys/django-test-plus

#django-debug-toolbar==1.9.1  # https://github.com/jazzband/django-debug-toolbar
django-extensions==3.2.3  # https://github.com/django-extensions/django-extensions
#django-coverage-plugin==1.5.0  # https://github.com/nedbat/django_coverage_plugin
#pytest-django==3.2.1  # https://github.com/pytest-dev/pytest-django
